+++ rts/c/backends/c.h
@@ -7,7 +7,7 @@ struct futhark_context_config {
   char logging;
   char tracing;
   char pedantic;
-  const char *cache_fname;
+  char *cache_fname;
   int num_tuning_params;
   int64_t *tuning_params;
   const char** tuning_param_names;
@@ -45,6 +45,7 @@ struct futhark_context {
   FILE *log;
   struct constants *constants;
   struct free_list free_list;
+  struct event_list event_list;
   int64_t peak_mem_usage_default;
   int64_t cur_mem_usage_default;
   struct program* program;