+++ rts/c/context.h
@@ -1,9 +1,5 @@
 // Start of context.h
 
-// Eventually it would be nice to move the context definition in here
-// instead of generating it in the compiler.  For now it defines
-// various helper functions that must be available.
-
 // Internal functions.
 
 static void set_error(struct futhark_context* ctx, char *error) {
@@ -80,6 +76,17 @@ static void host_free(struct futhark_context* ctx,
   }
 }
 
+static void add_event(struct futhark_context* ctx,
+                      const char* name,
+                      char* description,
+                      void* data,
+                      event_report_fn f) {
+  if (ctx->logging) {
+    fprintf(ctx->log, "Event: %s\n%s\n", name, description);
+  }
+  add_event_to_list(&ctx->event_list, name, description, data, f);
+}
+
 struct futhark_context_config* futhark_context_config_new(void) {
   struct futhark_context_config* cfg = malloc(sizeof(struct futhark_context_config));
   if (cfg == NULL) {
@@ -147,6 +154,7 @@ struct futhark_context* futhark_context_new(struct futhark_context_config* cfg) 
   //create_lock(&ctx->lock);
   if (cfg->tracing) printf("TRACE: rts: futhark_context_new: init free list...\n");
   free_list_init(&ctx->free_list);
+  event_list_init(&ctx->event_list);
   ctx->peak_mem_usage_default = 0;
   ctx->cur_mem_usage_default = 0;
   ctx->constants = malloc(sizeof(struct constants));
@@ -186,6 +194,8 @@ void futhark_context_free(struct futhark_context* ctx) {
   free_all_in_free_list(ctx);
   if (cfg->tracing) printf("TRACE: rts: futhark_context_free: destroy free list...\n");
   free_list_destroy(&ctx->free_list);
+  if (cfg->tracing) printf("TRACE: rts: futhark_context_free: free event list...\n");
+  event_list_free(&ctx->event_list);
   if (cfg->tracing) printf("TRACE: rts: futhark_context_free: free constants...\n");
   free(ctx->constants);
   if (cfg->tracing) printf("TRACE: rts: futhark_context_free: free error...\n");